requirements.txt
# ------------------------
# Core Backend
# ------------------------
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.1

# ------------------------
# Database (Supabase PostgreSQL)
# ------------------------
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# ------------------------
# Git & Project Handling
# ------------------------
GitPython==3.1.40

# ------------------------
# AI / LLM
# ------------------------
openai==1.3.7
transformers==4.36.2
torch>=2.0.0

# ------------------------
# Execution & Testing
# ------------------------
pytest==7.4.3

# ------------------------
# Utilities
# ------------------------
diff-match-patch==20230430
requests==2.31.0

# ------------------------
# Sandbox / Container (Optional but Recommended)
# ------------------------
docker==6.1.3
